% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{add_criticmarkup_theme}
\alias{add_criticmarkup_theme}
\title{Add CriticMarkup highlighting themes to RStudio}
\usage{
add_criticmarkup_theme(theme = "textmate", apply = FALSE,
  force = FALSE, globally = FALSE)
}
\arguments{
\item{theme}{Name of the CriticMarkup-enabled theme to apply.  Currently only
"textmate" is available}

\item{apply, force, globally}{passed to \code{\link[rstudioapi:addTheme]{rstudioapi::addTheme()}}}
}
\description{
Add CriticMarkup highlighting themes to RStudio
}
\examples{
\dontrun{
\donttest{
add_criticmarkup_theme("textmate")
}}
}
