% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{roundtrip_active_file}
\alias{roundtrip_active_file}
\alias{addins}
\alias{redoc_addins}
\alias{dedoc_to_active_file}
\alias{dedoc_to_new_file}
\title{RStudio Addin Functions}
\usage{
roundtrip_active_file()

dedoc_to_active_file(docx = NULL, showdiff = TRUE)

dedoc_to_new_file(docx = NULL, showdiff = TRUE)
}
\arguments{
\item{docx}{The input Word file, originally generated by \code{\link[=redoc]{redoc()}}}

\item{showdiff}{Display a diff of the current version of the document against
the rendered version?}
}
\description{
These functions act on the RStudio text editor and can be used via the
Addins menu or the R console. \code{rountrip_active_file()} ("Render and Update")
knits the current file using \code{\link[=redoc]{redoc()}} replaces the text with text that has
been round-tripped to Word and back. This helps eliminate small formatting
differences.

The \code{dedoc_*} functions de-render Word documents (using \code{\link[=dedoc]{dedoc()}}),
and place the results in the current or new file.  By default, they will
also display a diff (generated via \code{\link[=redoc_diff]{redoc_diff()}}) of edits in the word file
compared to its original version a compilation time. If called without
an input file, they will prompt for file selection via the RStudio GUI.
}
