% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticmarkup-addins.R
\name{cm_accept_all}
\alias{cm_accept_all}
\alias{cm_reject_all}
\alias{cm_delete_comments}
\title{Process CriticMarkup text}
\usage{
cm_accept_all(text)

cm_reject_all(text)

cm_delete_comments(text)
}
\arguments{
\item{text}{The text to process. Takes a character vector, but multi-line
text should be concatenated into a single character string with line
breaks if there are edits a ross lines.}
}
\value{
A character vector of the modified text.
}
\description{
These functions process CriticMarkup syntax in text and return the
modified text with changes accepted or rejected or comments deleted.
See \code{\link[=cm_process_file]{cm_process_file()}} to modify a file on-disk.
}
\examples{

text <- paste(readLines(redoc_example_criticmarkup()), collapse = '\\n')
cat(text)
cat(cm_accept_all(text))
cat(cm_reject_all(text))
cat(cm_delete_comments(cm_reject_all(text)))

}
