% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criticmarkup-addins.R
\name{cm_process_file}
\alias{cm_process_file}
\title{Accept or reject changes in CriticMarkup syntax in bulk}
\usage{
cm_process_file(file, action, out = file)
}
\arguments{
\item{file}{The name of the Markdown or R Markdown file}

\item{action}{One of "accept_all", "reject_all", or "delete_comments"}

\item{out}{The path to write the changed file to. The default is to
overwrite the original.}
}
\value{
The path of the modified file
}
\description{
This function modifies a Markdown or R Markdown file by accepting or rejecting
changes in CriticMarkup syntax. \link[=cm_accept_all]{Lower-level functions}
operate on text if the user wants to integrate these into a pipeline.
}
\examples{

outfile = tempfile(fileext = ".md")
cm_process_file(redoc_example_criticmarkup(), action = "accept_all", out = outfile)
cat(readLines(outfile), sep = "\\\\n")
}
