% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedoc.R
\name{dedoc}
\alias{dedoc}
\title{Convert an Reversible Document back to R Markdown}
\usage{
dedoc(docx, to = NULL, dir = ".", track_changes = "comments_only",
  block_missing = "comment", inline_missing = "omit",
  wrap = getOption("redoc.wrap", 80), overwrite = FALSE,
  orig_docx = NULL, orig_codefile = NULL, verbose = FALSE)
}
\arguments{
\item{docx}{The `.docx`` file to convert}

\item{to}{the filename to write the resulting \code{.Rmd} file.  The default is to
use the same basename as the docx document}

\item{dir}{The directory to write the `.Rmd`` to. Defaults to current working
directory}

\item{track_changes}{How to deal with tracked changes and comments in the
\code{.docx} file. \code{"accept"} accepts all changes, and \code{"reject"} rejects all of
them. The default, \code{"criticmarkup"}, converts the tracked changes to
\href{http://criticmarkup.com/spec.php#thebasicsyntax}{Critic Markup syntax}.
The default, \code{"comments_only"} will only convert comments, as other changes can be
viewed with \code{\link[=redoc_diff]{redoc_diff()}}.
\code{"all"} marks up tracked changes and comments in \code{<span>} tags and is
useful for debugging.  See the
\href{http://pandoc.org/MANUAL.html#option--track-changes}{pandoc manual} for
details.}

\item{block_missing, inline_missing}{What to do about code blocks or inline code
whose output has been removed in the editing of the Word document. "restore"
attempts to restore the code as close to its original location in the document
as possible. "comment" will do so but wrap it in HTML comments. "omit" will
not restore the code at all.}

\item{wrap}{The width at which to wrap text. If \code{NA}, text is not wrapped.
Set the default with \code{"redoc.wrap"} in \code{options()}.}

\item{overwrite}{Whether to overwrite an existing file}

\item{orig_codefile, orig_docx}{The original \code{.codelist.yml} or Word document
created when the document was first knit.  Useful for debugging, or in
cases where the word file has been corrupted or transformed, for instance,
by copy-and-pasting the content into a new file.  If provided, \code{dedoc} will
use this codefile or word file to re-create the \code{.Rmd} file with the text
of the input.}

\item{verbose}{whether to print pandoc progress text}
}
\description{
Converts a document originally created with \code{\link[=redoc]{redoc()}} back to R
Markdown, including changes made to text in MS Word.
}
\details{
R chunks may be lost in the editing process if using non-Microsoft
word processors (e.g. LibreOffice or in copy-and-pasting text into a new document.
}
