% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx-utils.R
\name{is_redoc}
\alias{is_redoc}
\title{Is this a reversible document?}
\usage{
is_redoc(docx)
}
\arguments{
\item{docx}{A path to a `.docx` file or an `rdocx` object produced by
[officer::read_docx()]}
}
\value{
a logical value
}
\description{
A function for testing is the file can be un-knit.  If not, un-knitting
may be attempted with the `orig_chunkfile` or `orig_docx` files in [undoc()].
}
\examples{
is_redoc(redoc_example_docx())
}
