% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessor.R
\name{make_preknitter}
\alias{make_preknitter}
\title{Generate a function to run prior to knitting}
\usage{
make_preknitter(wrappers = list())
}
\arguments{
\item{wrappers}{a list of \link[=make_wrapper]{wrapper functions}.}
}
\value{
A function to be used in `output_format(pre_knit = ...)``
}
\description{
\strong{redoc} modifies R Markdown documents prior to knitting to
capture and store document elements that will need to be restored later.  It
does this via a function passed to the \code{pre_knit} argument of
\code{\link[rmarkdown:output_format]{rmarkdown::output_format()}}.  \code{make_preknitter} generates this function from
a list of \link[=make_wrapper]{wrapper functions}.
}
\details{
\strong{rmarkdown} does not provide a mechanism to modify the file prior
to knitting, so \strong{redoc} accomplishes this by reaching up the call stack and
modifying the environment in the \code{\link[rmarkdown:render]{rmarkdown::render()}} function.  The
function generated by \code{make_preknitter} will do this by generating the
pre-processed R Markdown file and associated list of code chunks, adding these
the list of intermediates for cleanup, and switching the document input to
this pre-processed document.
}
\examples{

make_preknitter(wrappers = list(htmlcommentwrap, latexwrap))

}
