% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pandoc_ast}
\alias{pandoc_ast}
\title{Convert a document to Pandoc's abstract syntax tree format}
\usage{
pandoc_ast(file, from = NULL, tolist = TRUE)
}
\arguments{
\item{file}{the file to convert using pandoc.}

\item{from}{the format to convert from.  If \code{NULL} (default) File type will be
auto-detected by extension. \code{.Rmd} files will be treated as \code{.md}.}

\item{tolist}{whether to return the AST as an R list.  If \code{FALSE}, will
return length-1 character vector of raw JSON.}
}
\value{
A list containing the structured document
}
\description{
This is a convenience function for testing and development.
}
\examples{
ast <- pandoc_ast(redoc_example_docx())
}
