% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{parse_rmd_to_df}
\alias{parse_rmd_to_df}
\title{Extract the R and YAML chunks from an Rmd Document}
\usage{
parse_rmd_to_df(input_file)
}
\arguments{
\item{input_file}{An R Markdown File}
}
\value{
A data frame with fields \code{label}, \code{type}, \code{header}, and \code{code}
}
\description{
Extract the R and YAML chunks from an Rmd Document
}
\examples{
parse_rmd_to_df(redoc_example_rmd())
}
