% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedoc.R
\name{preprocess_docx}
\alias{preprocess_docx}
\title{Make modifications to docx xml before converting to markdown
Unzips to a temporary directory, modifies the XML, then re-zips
to a temporary file and returns its path
Maybe this should be refactored to use officer::read_docx and work
like other functions in officer, e.g., body_sdt_to_ins()}
\usage{
preprocess_docx(docx, outfile = tempfile(fileext = ".docx"))
}
\description{
Make modifications to docx xml before converting to markdown
Unzips to a temporary directory, modifies the XML, then re-zips
to a temporary file and returns its path
Maybe this should be refactored to use officer::read_docx and work
like other functions in officer, e.g., body_sdt_to_ins()
}
