% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx_reversible.R
\name{rdocx_reversible}
\alias{rdocx_reversible}
\title{Convert to a Reversible Microsoft Word Document}
\usage{
rdocx_reversible(highlight_outputs = FALSE, wrap = 80, ...)
}
\arguments{
\item{highlight_outputs}{whether to highlight outputs from chunks and inline
code in the final document}

\item{wrap}{when round-tripping the document, at what width to wrap the
markdown output? See [undoc()].}

\item{...}{other parameters passed to [rmarkdown::word_document()]}
}
\description{
Format for converting from R Markdown to a Microsoft Word Document that can
be reversed using [undoc()] after editing in Word.
}
