% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redoc-package.R
\name{redoc-package}
\alias{redoc-package}
\title{Reversible Reproducible Documents}
\description{
Implements a reversible 'R Markdown' to 'Microsoft Word'
pipeline.
}
\section{Support}{


The development repository for the \strong{redoc} package is found at
\url{https://github.com/noamross/redoc}.  Please file
bug reports or other feedback at
\url{https://github.com/noamross/redoc/issues}
}

\author{
Noam Ross \email{noam.ross@gmail.com}
}
\keyword{package}
