% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redoc.R
\name{redoc}
\alias{redoc}
\title{R Markdown format for Reversible Reproducible Word Documents}
\usage{
redoc(highlight_outputs = FALSE, wrap = getOption("redoc.wrap", 80),
  margins = NULL, line_numbers = NULL, comment_author = NULL,
  keep_md = FALSE, wrappers = list(htmlcommentwrap, latexwrap,
  rawblockwrap, rawspanwrap, cmwrap, citationwrap), diagnostics = TRUE,
  ...)
}
\arguments{
\item{highlight_outputs}{whether to highlight outputs from chunks and inline
code in the final document}

\item{wrap}{when round-tripping the document, at what width to wrap the
markdown output? Set the default with \code{"redoc.wrap"} in \code{options()}. See \code{\link[=dedoc]{dedoc()}}.}

\item{margins}{page margin size.  Can be a single value or a named vector
with values, \code{top}, \code{bottom}, \code{left}, \code{right}, \code{gutter}, \code{header}, and
\code{footer}.  If NULL defaults to the reference document.}

\item{line_numbers}{either TRUE or list with any of the arguments \code{start},
\code{by}, \code{restart}, and \code{distance}}

\item{comment_author}{The name to attribute any CriticMarkup tracked
changes to. Defaults to \code{\link[whoami:fullname]{whoami::fullname()}}.}

\item{keep_md}{whether to keep the intermediate markdown document}

\item{wrappers}{a list of wrapper functions to capture text to protect when
rendering and de-rendering.  See \code{\link[=make_wrapper]{make_wrapper()}}.}

\item{diagnostics}{Whether to embed diagnostic information in the output
file.  If TRUE, this will save session information and the current
pandoc and (if used) RStudio versions inside the Word document for later
bug-checking.}

\item{...}{other parameters passed to \code{\link[rmarkdown:word_document]{rmarkdown::word_document()}}}
}
\description{
Format for converting from R Markdown to a Microsoft Word Document that can
be reversed using \code{\link[=dedoc]{dedoc()}} after editing in Word.
}
