% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx-utils.R
\name{redoc_example_docx}
\alias{redoc_example_docx}
\title{Path to an example Revserible Microsoft Word file}
\usage{
redoc_example_docx()
}
\description{
Path to an example Revserible Microsoft Word file
}
\examples{
redoc_example_docx()
}
