% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docx-utils.R
\name{redoc_example_rmd}
\alias{redoc_example_rmd}
\alias{redoc_examples}
\alias{redoc_example_docx}
\alias{redoc_example_edited_docx}
\alias{redoc_example_criticmarkup}
\title{Files for examples and testing}
\usage{
redoc_example_rmd()

redoc_example_docx()

redoc_example_edited_docx()

redoc_example_criticmarkup()
}
\description{
Files for examples and testing
}
\examples{
redoc_example_rmd()
redoc_example_docx()
redoc_example_edited_docx()
}
