% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dedoc.R
\name{redoc_extract_rmd}
\alias{redoc_extract_rmd}
\title{Extract the Rmd used to to produce a Reversible Word Doc}
\usage{
redoc_extract_rmd(docx, type = c("original", "roundtrip"), dir = ".",
  to = NULL, overwrite = FALSE)
}
\arguments{
\item{docx}{A path to a word file or a an \code{rdocx} object created with
\code{\link[officer:read_docx]{officer::read_docx()}}.}

\item{type}{One of \code{"original"} or \code{"roundtrip"}. \code{"original"} extracts the
exact document originally knit.  \code{"roundtrip"} (default) extracts a document
that has been converted to Word and back with no edits in between.  The
latter should be more useful for comparing against edits, as line-wrapping
and placement of no-output chunks should match.}

\item{dir}{The directory to write the `.Rmd`` to. Defaults to current working
directory}

\item{to}{the filename to write the resulting \code{.Rmd} file.  The default is to
use the the original name with either \code{.orignal.Rmd} or \code{roundtrip.Rmd}
extensions.}

\item{overwrite}{whether to overwrite existing files}
}
\value{
The path to the extracted \code{.Rmd}
}
\description{
Documents produced with \code{\link[=redoc]{redoc()}} store an copy of the original
\code{.Rmd} files used to produce them.  This is useful for diffing against the
version created with \code{\link[=dedoc]{dedoc()}}, especially if tracked changes have not been
used.
}
\examples{
redoc_extract_rmd(redoc_example_docx(), dir = tempdir())
}
