% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{redoc_index}
\alias{redoc_index}
\title{Path to Redoc Index}
\usage{
redoc_index()
}
\description{
Retrieves the path to the redoc index file.
}
\examples{
\dontrun{
if (interactive()) {
  browseURL(redoc_index())
} else {
  print(paste("You can use redoc under: ", redoc_index()))
}
}
}
