% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{redoc_path}
\alias{redoc_path}
\title{Path to Redoc Resources}
\usage{
redoc_path()
}
\description{
Retrieves the path to redoc resources.
}
\examples{
\dontrun{
if (interactive()) {
  browseURL(redoc_path())
} else {
  print(paste("You can explore redoc resources under: ", redoc_path()))
}
}
}
