% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{redoc_spec}
\alias{redoc_spec}
\title{Redoc Index File with OpenAPI Path}
\usage{
redoc_spec(spec_url = "https://redocly.github.io/redoc/openapi.yaml", ...)
}
\arguments{
\item{spec_url}{Url to an openAPI specification}

\item{...}{Additional options for Redoc. See
\url{https://github.com/Redocly/redoc#redoc-options-object}}
}
\value{
large string containing the contents of \code{\link{redoc_index}()} with
  the appropriate speicification path changed to the \code{spec_url} value.
}
\description{
Produces the content for a \code{index.html} file that will attempt to access a
provided OpenAPI Specification URL.
}
\examples{
\dontrun{
if (interactive()) {
  redoc_spec("https://docs.docker.com/engine/api/v1.38.yaml",
             scrollYOffset = 250,
             disableSearch = TRUE)
}
}
}
