% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redoc.R
\docType{package}
\name{redoc-package}
\alias{redoc}
\alias{redoc-package}
\title{redoc: Generates Redoc documentation from an OAS Compliant API}
\description{
Redoc is a collection of HTML, JavaScript, CSS and fonts assets that
that generate Redoc documentation from an OAS compliant API.
}
\details{
The main purpose of this package is to enable package authors to create
APIs that are compatible with \url{https://redoc.ly/redoc/}
and \url{https://www.openapis.org/}.

To learn more about Redoc visit: \url{https://redoc.ly/redoc/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/meztez/redoc}
  \item Report bugs at \url{https://github.com/meztez/redoc/issues}
}

}
\author{
\strong{Maintainer}: Bruno Tremblay \email{openr@neoxone.com}

Authors:
\itemize{
  \item  Rebilly [copyright holder]
}

Other contributors:
\itemize{
  \item  The Montserrat Project Authors (https://github.com/JulietaUla/Montserrat) [copyright holder]
  \item  Google Inc. (https://fonts.google.com) [copyright holder]
}

}
