% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{redoc_extract_rmd}
\alias{redoc_extract_rmd}
\title{Extract the Rmd used to to produce a Reversible Word Doc}
\usage{
redoc_extract_rmd(docx, type = c("original", "roundtrip"), dir = ".",
  to = NULL, overwrite = FALSE)
}
\arguments{
\item{docx}{A path to a word file or a an `rdocx` object created with
[officer::read_docx()].}

\item{type}{One of `"original"` or `"roundtrip"`..  `"original"` extracts the
exact document originally knit.  `"roundtrip` (default) extracts a document
that has been converted to Word and back with no edits in between.  The
latter should be more useful for comparing against edits, as line-wrapping
and placement of no-output chunks should match.}

\item{dir}{The directory to write the `.Rmd`` to. Defaults to current working
directory}

\item{to}{the filename to write the resulting `.Rmd` file.  The default is to
use the the original name with either `.orignal.Rmd` or `roundtrip.Rmd`
extensions.}

\item{overwrite}{whether to overwite existing files}
}
\value{
The path to the extracted `.Rmd`
}
\description{
Documents produced with [docx_reversible()] store an copy of the original
`.Rmd` files used to produce them.  This is useful for diffing against the
version created with [undoc()], especially if tracked changes have not been
used.
}
\examples{
redoc_extract_rmd(redoc_example_docx(), dir = tempdir())
}
