% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{undoc}
\alias{undoc}
\title{Convert an Reversible Document back to R Markdown}
\usage{
undoc(docx, to = NULL, dir = ".", track_changes = c("criticmarkup",
  "accept", "reject", "all"), wrap = 80, overwrite = FALSE,
  orig_chunkfile = NULL, orig_docx = NULL, verbose = FALSE)
}
\arguments{
\item{docx}{The \code{.docx file to convert}}

\item{to}{the filename to write the resulting \code{.Rmd} file.  The default is to
use the same basename as the docx document}

\item{dir}{The directory to write the `.Rmd`` to. Defaults to current working
directory}

\item{track_changes}{How to deal with tracked changes and comments in the
\code{.docx} file. \code{"accept"} accepts all changes, and \code{"reject"} rejects all of
them. The default, \code{"criticmarkup"}, converts the tracked changes to
\href{http://criticmarkup.com/spec.php#thebasicsyntax}{Critic Markup syntax}.
\code{"all"} marks up tracked changes and comments in \code{<span>} tags.  See the
\href{http://pandoc.org/MANUAL.html#option--track-changes}{pandoc manual} for
details.}

\item{wrap}{The width at which to wrap text. If \code{NA}, text is not wrapped}

\item{overwrite}{Whether to overwrite an existing file}

\item{orig_chunkfile, orig_docx}{The original chunkfile or Word document
created when the document was first knit.  Useful for debugging, or in
cases where the word file has been corrupted or transformed, for instance
by copy-and-pasting the content into a new file.  If provided, undoc will
use this chunkfile or word file to re-create the \code{.Rmd} file with the text
of the input.}

\item{verbose}{whether to print pandoc progress text}
}
\description{
Converts a document originally created with \code{\link[=rdocx_reversible]{rdocx_reversible()}} back to R
Markdown, including changes made to text in MS Word.
}
\details{
Internal data storing R chunks may be lost in the editing process.
This is known to occur with documents edited with LibreOffice.
}
